refactor(TP-162): delete task-runner.ts and clean up test dependencies (Steps 1-3)
