+++ taskplane-tasks/TP-162-task-runner-delete-and-cleanup/.reviews/R004-plan-step3.md
@@ -1,0 +1,15 @@
+## Plan Review: Step 3: Delete task-runner.ts
+
+### Verdict: APPROVE
+
+### Summary
+This revised Step 3 plan is now outcome-complete for safely deleting `extensions/task-runner.ts` without breaking the test suite. It addresses the previously flagged gaps by explicitly handling remaining test imports, source-reading tests, and residual describe blocks that hard-reference `task-runner.ts`. The added final reference check gives a clear guardrail before deletion.
+
+### Issues Found
+1. **[Severity: minor]** — No blocking issues found for Step 3 execution.
+
+### Missing Items
+- None blocking for this step.
+
+### Suggestions
+- Consider adding one explicit grep checkpoint for non-test references immediately after deletion (for example, `extensions/tsconfig.json` currently includes `"task-runner.ts"`). This is likely handled in Step 4’s “additional files from grep” bucket, but making it explicit can reduce cleanup drift.
